Convert to ES modules, bump min node to 16.0.0
