+++ bin/cli.mjs
@@ -2,10 +2,11 @@
 
 /* eslint-disable no-console */
 
-const chalk = require('chalk')
-const program = require('commander')
-const {name, version} = require('../package.json')
-const {createBumpr} = require('../src/cli')
+import chalk from 'chalk'
+import program from 'commander'
+import createBumpr from '../src/cli.mjs'
+
+import {name, version} from '../package.mjs'
 
 function handleError(error) {
   const msg = error.message ? error.message : error
