--- jest.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  collectCoverageFrom: ['src/**/*.js', '!src/typedefs.js', '!src/**/tests/*.js'],
-  coverageThreshold: {
-    global: {
-      branches: 100,
-      statements: 100
-    }
-  },
-  testEnvironment: 'node',
-  testMatch: ['<rootDir>/**/tests/*.test.js']
-}
