+++ package.mjs
@@ -1,0 +1,4 @@
+import {createRequire} from 'module'
+
+const require = createRequire(import.meta.url)
+export const {name, version} = require('./package.json')
