--- src/ci/circle.js
@@ -1,13 +1,0 @@
-require('../typedefs')
-
-const CiBase = require('./base')
-
-/**
- * CI interface for public Circle (circleci.com)
- *
- * @class
- * @implements {Ci}
- */
-class Circle extends CiBase {}
-
-module.exports = Circle
