+++ src/ci/circle.mjs
@@ -1,0 +1,11 @@
+import '../typedefs.mjs'
+
+import CiBase from './base.mjs'
+
+/**
+ * CI interface for public Circle (circleci.com)
+ *
+ * @class
+ * @implements {Ci}
+ */
+export default class Circle extends CiBase {}
