+++ src/errors/no-log-file.mjs
@@ -1,0 +1,7 @@
+import chalk from 'chalk'
+
+export default class NoLogFileError extends Error {
+  constructor(logFile) {
+    super(`log file ${chalk.magentaBright(logFile)} not found.`)
+  }
+}
