+++ src/vcs/github.mjs
@@ -1,11 +1,11 @@
-require('../typedefs')
+import '../typedefs.mjs'
 
-const {readFileSync} = require('fs')
-const fetch = require('node-fetch')
-const path = require('path')
+import {readFileSync} from 'fs'
+import fetch from 'node-fetch'
+import path from 'path'
 
-const {Logger} = require('../logger')
-const {exec} = require('../node-wrappers')
+import Logger from '../logger.mjs'
+import {exec} from '../node-wrappers.mjs'
 
 /**
  * Get fetch options
@@ -251,4 +251,4 @@ class GitHub {
   }
 }
 
-module.exports = GitHub
+export default GitHub
